Make primitive offset table optional (#165)

+++ akka-projection-r2dbc/src/test/scala/akka/projection/r2dbc/TestDbLifecycle.scala
@@ -42,10 +42,12 @@ trait TestDbLifecycle extends BeforeAndAfterAll { this: Suite =>
       r2dbcExecutor.updateOne("beforeAll delete")(
         _.createStatement(s"delete from ${r2dbcSettings.durableStateTableWithSchema}")),
       10.seconds)
-    Await.result(
-      r2dbcExecutor.updateOne("beforeAll delete")(
-        _.createStatement(s"delete from ${r2dbcProjectionSettings.offsetTableWithSchema}")),
-      10.seconds)
+    if (r2dbcProjectionSettings.isOffsetTableDefined) {
+      Await.result(
+        r2dbcExecutor.updateOne("beforeAll delete")(
+          _.createStatement(s"delete from ${r2dbcProjectionSettings.offsetTableWithSchema}")),
+        10.seconds)
+    }
     Await.result(
       r2dbcExecutor.updateOne("beforeAll delete")(
         _.createStatement(s"delete from ${r2dbcProjectionSettings.timestampOffsetTableWithSchema}")),
